% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_reg_predict.R
\name{predict_class._glmnetcr}
\alias{predict_class._glmnetcr}
\title{These are internal functions not meant to be directly called by the user.}
\usage{
predict_class._glmnetcr(object, new_data, ...)
}
\arguments{
\item{object}{fitted model}

\item{new_data}{new data}

\item{...}{other arguments}
}
\description{
These are internal functions not meant to be directly called by the user.
}
