% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_imports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{multi_predict}
\alias{predict_numeric}
\alias{predict_raw}
\alias{translate}
\alias{min_grid}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{parsnip}{\code{\link[parsnip]{multi_predict}}, \code{\link[parsnip:other_predict]{predict_numeric}}, \code{\link[parsnip:predict.model_fit]{predict_raw}}, \code{\link[parsnip]{translate}}}

  \item{tune}{\code{\link[tune:reexports]{min_grid}}}
}}

