% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_earth.R
\name{required_pkgs.step_earth}
\alias{required_pkgs.step_earth}
\title{S3 methods for tracking which additional packages are needed for steps.}
\usage{
required_pkgs.step_earth(x, ...)
}
\arguments{
\item{x}{A recipe step}

\item{...}{Additional arguments}
}
\value{
A character vector
}
\description{
S3 methods for tracking which additional packages are needed for steps.
}
