% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_earth.R
\name{step_earth}
\alias{step_earth}
\title{Multivariate Adaptive Regression Splines}
\usage{
step_earth(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  drop = TRUE,
  outcome = NULL,
  options = list(),
  res = NULL,
  prefix = "EARTH",
  skip = FALSE,
  id = recipes::rand_id("earth")
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence
of operations for this recipe.}

\item{...}{One or more selector functions to choose variables.}

\item{role}{For model terms created by this step, what analysis role
should they be assigned?. By default, the function assumes that
the new columns created from the original variables will be used
as predictors in a model.}

\item{trained}{A logical to indicate if the quantities
for preprocessing have been estimated.}

\item{drop}{drop}

\item{outcome}{Outcome column string.}

\item{options}{List of options to \code{earth::earth}.}

\item{res}{A list of results are stored here once this preprocessing step
has been trained by \link[recipes:prep]{prep.recipe()}.}

\item{prefix}{A character string that will be the prefix to the resulting
new variables.}

\item{skip}{A logical. Should the step be skipped when the
recipe is baked by \code{\link[=bake.recipe]{bake.recipe()}}? While all operations are baked
when \code{\link[=prep.recipe]{prep.recipe()}} is run, some operations may not be able to be
conducted on new data (e.g. processing the outcome variable(s)).
Care should be taken when using \code{skip = TRUE} as it may affect
the computations for subsequent operations}

\item{id}{A character string that is unique to this step to identify it.}
}
\description{
\code{step_earth} creates a \emph{specification} of a recipe step that will create
new columns that are Multivariate Adaptive Regression Splines
based on \link[earth:earth]{earth} model.
}
