% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_glmnet.R
\name{weighted_glmnet}
\alias{weighted_glmnet}
\title{Weighted linear model}
\usage{
weighted_glmnet(x, y, ..., weight_column = "weights")
}
\arguments{
\item{x}{predictors matrix}

\item{y}{target vector}

\item{...}{passed to \code{glmnet}}

\item{weight_column}{name of weights column in \code{x}}
}
\value{
'glmnet' object
}
\description{
Weighted linear model
}
